requirements.txt
# requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0